# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GraphNode API
  version: 0.1.0
  description: >-
    Backend API for GraphNode. Errors follow RFC 9457 Problem Details.
...
paths:
...
  /v1/ai/conversations/{conversationId}:
    get:
      summary: Get conversation
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: ../schemas/conversation.json }
              examples:
                default:
                  externalValue: ./examples/ai-conversation-get-200.json
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema: { $ref: ../schemas/problem.json }
    patch:
      summary: Update conversation
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
            examples:
              default:
                externalValue: ./examples/ai-conversation-update-request.json
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: ../schemas/conversation.json }
              examples:
                default:
                  externalValue: ./examples/ai-conversation-update-200.json
    delete:
      summary: Delete conversation
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
        - in: query
          name: permanent
          schema: { type: boolean, default: false }
          description: If true, performs a hard delete. Otherwise, soft delete.
      responses:
        '204':
          description: No Content
...